trying io binding

+++ distilbert-finetune/inference.py
@@ -67,7 +67,7 @@ def infer(args):
             name='input_ids',
             device_type='cuda',
             device_id=0,
-            element_type=np.float32,
+            element_type=np.int64,
             shape=tuple(input_ids_tensor.shape),
             buffer_ptr=input_ids_tensor.data_ptr(),
         )
